--- a/EDDiscovery/UserControls/UserControlScan.cs
+++ b/EDDiscovery/UserControls/UserControlScan.cs
@@ -395,7 +395,7 @@ namespace EDDiscovery.UserControls
                         labelvoff -= 15;
                     }
 
-                    endpoint = CreateImageLabel(pc, sc.GetStarTypeImage().Item1,
+                    endpoint = CreateImageLabel(pc, sc.GetStarTypeImage(),
                                                 new Point(curpos.X + offset, curpos.Y + alignv),      // WE are basing it on a 1/4 + 1 + 1/4 grid, this is not being made bigger, move off
                                                 size, starLabel, tip, alignv + labelvoff, sc.IsEDSMBody, false);          // and the label needs to be a quarter height below it..
 
@@ -431,7 +431,7 @@ namespace EDDiscovery.UserControls
                 {
                     bool indicatematerials = sc.HasMaterials && !checkBoxMaterials.Checked;
 
-                    Image nodeimage = sc.IsStar ? sc.GetStarTypeImage().Item1 : sc.GetPlanetClassImage();
+                    Image nodeimage = sc.IsStar ? sc.GetStarTypeImage() : sc.GetPlanetClassImage();
 
                     if (ImageRequiresAnOverlay(sc, indicatematerials))
                     {