* Cache name to type ID lookup (helps with larger scan histories)
* Split GetStarTypeImage into two functions (cuts about 1ms per star scan)
* Use TypeOfJournalEntry(JournalTypeEnum) when possible to get journal event type (cuts about 10ms per 1000 journal entries)
* Use the JournalFriends object and not its json (down to 2ms from 175ms)
* Disable EDSM fetch from FindEDSM